# apps/backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Link Shortener API",
    "description": "Api for Link Shortener application",
    "version": "1.0.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format **Bearer &lt;token>**"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "tags": [
    {
      "name": "auth"
    },
    {
      "name": "link"
    },
    {
      "name": "link-redirect"
    }
  ],
  "paths": {
    "/ping": {
      "get": {
        "description": "Just for testing",
        "responses": {
          "200": {
            "description": "Successfully send a message back",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user",
        "description": "This endpoint registers a new user with a username, email, and password.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User registration data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRegistration"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login a user",
        "description": "This endpoint authenticates a user by email and password.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "schema": {
              "$ref": "#/definitions/UserToken"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/link/shorten": {
      "post": {
        "tags": ["link"],
        "summary": "Shorten the link",
        "description": "This endpoint shorten the link",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Data required for shorten the link",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LinkCreateSchemaType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link shorten successfully",
            "schema": {
              "$ref": "#/definitions/IShortUrl"
            }
          },
          "400": {
            "description": "A shortened link already exists for this user with the same original URL or input data invalid"
          }
        }
      }
    },
    "/link/{shortId}": {
      "get": {
        "tags": ["link"],
        "summary": "Retrieve the link",
        "description": "This returns the link",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "shortId",
            "description": "Shorten link id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Link shorten",
            "schema": {
              "$ref": "#/definitions/IShortUrl"
            }
          },
          "404": {
            "description": "Link not found"
          }
        }
      },
      "delete": {
        "tags": ["link"],
        "summary": "Delete the link",
        "description": "This endpoint deletes the link",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "shortId",
            "description": "Shorten link id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Link shorten"
          },
          "404": {
            "description": "Link not found"
          }
        }
      }
    },
    "/{shortId}": {
      "get": {
        "tags": ["link-redirect"],
        "summary": "Link Redirect",
        "description": "Redirect to the link",
        "parameters": [
          {
            "in": "path",
            "name": "shortId",
            "description": "Shorten link id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Link shorten"
          },
          "404": {
            "description": "Link not found"
          }
        }
      }
    }
  },
  "definitions": {
    "UserRegistration": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "email",
        "password"
      ]
    },
    "UserLogin": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "UserToken": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "LinkCreateSchemaType": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string"
        }
      }
    },
    "IShortUrl": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string"
        },
        "shortId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        }
      }
    }
  }
}